backend/requirements.txt
absl-py==1.0.0
asgiref==3.5.0
astunparse==1.6.3
backports.zoneinfo==0.2.1
boto3==1.21.2
botocore==1.24.2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cmake==3.22.2
colormath==3.0.0
cycler==0.11.0
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
django-hashid-field==3.3.4
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
fonttools==4.29.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.1
imutils==0.5.4
jmespath==0.10.0
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
networkx==2.6.3
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.1
scikit-learn==0.22.1
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
typing-extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
